Move run function to separate file
This is mainly to avoid unwanted imports prior to starting the server
